Adding options and supporting read-file options
Adding a suggestions translation blacklist and a checkbox to switch between it using strings or regexes. The blacklist is loaded from a file, so I added a widget for selecting files to read. There was already one for selecting files for writing to, but that would show a warning if the file already existed, which in this case we expect it to do.